+++ inst/chunking/Autosubmit/startR_autosubmit.sh
@@ -11,6 +11,7 @@ chunknum="${chunknum:5}"  #e.g., CHUNK_0
 # e.g., chunk_args=(%JOBS."${chunknum}".dat% %JOBS."${chunknum}".dat_N% %JOBS."${chunknum}".var% %JOBS."${chunknum}".var_N% %JOBS."${chunknum}".sdate% %JOBS."${chunknum}".sdate_N%)
 chunk_args=
 
+include_special_setup
 include_init_commands
 include_module_load
 
@@ -21,5 +22,5 @@ include_module_load
 cd_run_dir
 
 #e.g., Rscript load_process_save_chunk_autosubmit.R --args $task_path 1 1 1 1 2 2 1 1 1 2 1 2
-Rscript ${proj_dir}/load_process_save_chunk_autosubmit.R --args ${task_path} ${chunk_args[@]}
+Rscript load_process_save_chunk_autosubmit.R --args ${task_path} ${chunk_args[@]}